requirements.txt
fastapi
uvicorn
openai
requests
pydantic
pillow
markdown
pandas
python-dotenv
